Move everything to a Java subpackage
