+++ firestore/app/src/androidTest/java/com/google/firebase/example/fireeats/MainActivityTest.java
@@ -5,11 +5,14 @@ import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.support.test.uiautomator.UiDevice;
 import android.support.test.uiautomator.UiObject;
-import android.support.test.uiautomator.UiSelector;
 import android.support.test.uiautomator.UiScrollable;
+import android.support.test.uiautomator.UiSelector;
 import android.test.suitebuilder.annotation.LargeTest;
 import android.view.accessibility.AccessibilityWindowInfo;
+
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.example.fireeats.java.MainActivity;
+
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
