+++ firestore/app/src/main/java/com/google/firebase/example/fireeats/java/FilterDialogFragment.java
@@ -1,4 +1,4 @@
-package com.google.firebase.example.fireeats;
+package com.google.firebase.example.fireeats.java;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -9,7 +9,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Spinner;
 
-import com.google.firebase.example.fireeats.model.Restaurant;
+import com.google.firebase.example.fireeats.R;
+import com.google.firebase.example.fireeats.java.model.Restaurant;
 import com.google.firebase.firestore.Query;
 
 import butterknife.BindView;
