+++ firestore/app/src/main/java/com/google/firebase/example/fireeats/java/model/Restaurant.java
@@ -1,4 +1,4 @@
-package com.google.firebase.example.fireeats.model;
+package com.google.firebase.example.fireeats.java.model;
 
 import com.google.firebase.firestore.IgnoreExtraProperties;
